Created moneyrewards enum, and edited a bit

+++ src/com/dealornodeal/Briefcase.java
@@ -20,7 +20,7 @@ class Briefcase extends MoneyBoard{
         setNumber(number);
     }
 
-    public int getDollarAmount() {
+    public double getDollarAmount() {
         return dollarAmount;
     }
